+++ tasks/src/lib.rs
@@ -1,6 +1,4 @@
 extern crate failure;
-#[macro_use]
-extern crate failure_derive;
 extern crate futures;
 extern crate humthreads;
 #[macro_use]
@@ -15,6 +13,7 @@ extern crate serde_json;
 #[macro_use]
 extern crate slog;
 
+extern crate replicante_util_failure;
 extern crate replicante_util_rndid;
 
 
@@ -37,8 +36,9 @@ mod worker;
 
 pub use self::admin::TasksAdmin as Admin;
 pub use self::config::Config;
+pub use self::error::Error;
+pub use self::error::ErrorKind;
 pub use self::error::Result;
-pub use self::error::TaskError;
 pub use self::request::TaskRequest;
 pub use self::request::Tasks;
 pub use self::task_id::TaskId;