+++ tasks/src/request/backend/mock.rs
@@ -1,14 +1,15 @@
 use std::sync::Arc;
 use std::sync::Mutex;
 
+use failure::ResultExt;
 use serde_json;
 use serde_json::Value;
 
+use super::super::super::ErrorKind;
+use super::super::super::Result;
+use super::super::super::TaskQueue;
+use super::super::TaskRequest;
 use super::Backend;
-use super::Result;
-use super::TaskQueue;
-use super::TaskRequest;
-
 
 /// Mock implementation of a tasks queue backend.
 pub struct Mock<Q: TaskQueue> {
@@ -17,7 +18,8 @@ pub struct Mock<Q: TaskQueue> {
 
 impl<Q: TaskQueue> Backend<Q> for Mock<Q> {
     fn request(&self, task: TaskRequest<Q>, message: &[u8]) -> Result<()> {
-        let message: Value = serde_json::from_slice(message)?;
+        let message: Value = serde_json::from_slice(message)
+            .with_context(|_| ErrorKind::PayloadDeserialize)?;
         self.requests.lock().expect("failed to lock Mock#requests").push((task, message));
         Ok(())
     }