+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - A better thread story.
 - Introduce an `/api/unstable` API "version".
+- Threads introspection API.
 
 ### Changed
 - **BREAKING**: Rename incorrectly named v1 as unstable.